+++ booksim2/arbiters/matrix_arb.hpp
@@ -63,7 +63,8 @@ namespace Booksim
 
       // Arbitrate amongst requests. Returns winning input and 
       // updates pointers to metadata when valid pointers are passed
-      virtual int Arbitrate( int* id = 0, int* pri = 0) ;
+      //BSMOD: Change flit and packet id to long
+      virtual int Arbitrate( long* id = 0, int* pri = 0) ;
 
       virtual void AddRequest( int input, int id, int pri ) ;
